DemandSniper/requirements.txt
# FastAPI & Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
alembic>=1.12.1

# Web Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.2
httpx>=0.25.2
aiohttp>=3.9.1
fake-useragent>=1.4.0

# Data Processing
pandas>=2.1.3
numpy>=1.26.2

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Scheduling
apscheduler>=3.10.4

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
rapidfuzz>=3.5.2
tenacity>=8.2.3
dateparser>=1.1.8

# Email/Notifications
requests>=2.31.0
aiosmtplib>=3.0.1
jinja2>=3.1.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
